+++ tests/test_streams.py
@@ -41,8 +41,7 @@ class StreamsTest(unittest.TestCase):
     @mock.patch("flumine.streams.streams.Streams.add_historical_stream")
     def test_call_backtest(self, mock_add_historical_stream):
         self.mock_flumine.BACKTEST = True
-        mock_strategy = mock.Mock()
-        mock_strategy.streams = []
+        mock_strategy = mock.Mock(streams=[], historic_stream_ids=[])
         mock_strategy.market_filter = {
             "markets": ["dubs of the mad skint and british"],
             "listener_kwargs": {"canary_yellow": True},
@@ -53,6 +52,7 @@ class StreamsTest(unittest.TestCase):
             mock_strategy, "dubs of the mad skint and british", canary_yellow=True
         )
         self.assertEqual(len(mock_strategy.streams), 1)
+        self.assertEqual(len(mock_strategy.historic_stream_ids), 1)
 
     def test_call_backtest_no_markets(self):
         self.mock_flumine.BACKTEST = True
@@ -437,7 +437,7 @@ class TestHistoricalStream(unittest.TestCase):
     def test_handle_output(self):
         self.stream.handle_output()
 
-    @mock.patch("flumine.streams.historicalstream.HistoricalGeneratorStream")
+    @mock.patch("flumine.streams.historicalstream.FlumineHistoricalGeneratorStream")
     def test_create_generator(self, mock_generator):
         generator = self.stream.create_generator()
         mock_generator.assert_called_with(
@@ -483,15 +483,15 @@ class TestFlumineMarketStream(unittest.TestCase):
             mock.Mock(inplay=True, seconds_to_start=None), 0
         )
         self.stream._caches = {
-            "1.123": mock.Mock(market_definition={"status": "OPEN", "inPlay": False})
+            "1.123": mock.Mock(_definition_status="OPEN", _definition_in_play=False),
         }
         self.assertEqual(len(self.stream.snap()), 0)
         self.stream._caches = {
-            "1.123": mock.Mock(market_definition={"status": "OPEN", "inPlay": True})
+            "1.123": mock.Mock(_definition_status="OPEN", _definition_in_play=True),
         }
         self.assertEqual(len(self.stream.snap()), 1)
         self.stream._caches = {
-            "1.123": mock.Mock(market_definition={"status": "CLOSED", "inPlay": False})
+            "1.123": mock.Mock(_definition_status="CLOSED", _definition_in_play=False),
         }
         self.assertEqual(len(self.stream.snap()), 1)
 
@@ -499,11 +499,11 @@ class TestFlumineMarketStream(unittest.TestCase):
             mock.Mock(inplay=False, seconds_to_start=None), 0
         )
         self.stream._caches = {
-            "1.123": mock.Mock(market_definition={"status": "OPEN", "inPlay": False})
+            "1.123": mock.Mock(_definition_status="OPEN", _definition_in_play=False),
         }
         self.assertEqual(len(self.stream.snap()), 1)
         self.stream._caches = {
-            "1.123": mock.Mock(market_definition={"status": "OPEN", "inPlay": True})
+            "1.123": mock.Mock(_definition_status="OPEN", _definition_in_play=True),
         }
         self.assertEqual(len(self.stream.snap()), 0)
 
@@ -587,10 +587,10 @@ class TestHistoricListener(unittest.TestCase):
         )
 
     def test_on_data(self):
-        mock_stream = mock.Mock()
+        mock_stream = mock.Mock(_lookup="mc")
         self.listener.stream = mock_stream
-        self.listener.on_data("{}")
-        mock_stream.on_update.assert_called_with({})
+        self.listener.on_data('{"pt": 123, "mc": {}}')
+        self.listener.stream._process.assert_called_with({}, 123)
         # error
         self.assertIsNone(self.listener.on_data("p"))