+++ HISTORY.rst
@@ -3,6 +3,15 @@
 Release History
 ---------------
 
+1.16.1 (2021-01-28)
++++++++++++++++++++
+
+**Improvements**
+
+- Various optimisations on pending_packages and low level listener updates
+- Cache stream_id when backtesting
+- Always run integrations tests (now possible with faster backtesting from bflw 2.12.0)
+
 1.16.0 (2021-01-25)
 +++++++++++++++++++